Removed redundant enum definition within event_report_pdu file
